extract CallToActionViewHolder from CallToAction model
