+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/CallToActionViewHolder.java
@@ -1,0 +1,81 @@
+package org.cru.godtools.tract.model;
+
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.view.ViewCompat;
+import android.view.View;
+import android.widget.ImageView;
+import android.widget.TextView;
+
+import org.cru.godtools.tract.R;
+import org.cru.godtools.tract.R2;
+import org.cru.godtools.tract.util.DrawableUtils;
+
+import butterknife.BindView;
+import butterknife.OnClick;
+
+final class CallToActionViewHolder extends BaseViewHolder<CallToAction> {
+    public interface Callbacks {
+        void goToNextPage();
+    }
+
+    @BindView(R2.id.call_to_action_label)
+    TextView mLabelView;
+    @BindView(R2.id.call_to_action_arrow)
+    ImageView mArrowView;
+
+    @Nullable
+    private Callbacks mCallbacks;
+
+    CallToActionViewHolder(@NonNull final View root, @Nullable final BaseViewHolder parentViewHolder) {
+        super(CallToAction.class, root, parentViewHolder);
+    }
+
+    @NonNull
+    public static CallToActionViewHolder forView(@NonNull final View root,
+                                                 @Nullable final PageViewHolder parentViewHolder) {
+        final CallToActionViewHolder holder = forView(root, CallToActionViewHolder.class);
+        return holder != null ? holder : new CallToActionViewHolder(root, parentViewHolder);
+    }
+
+    /* BEGIN lifecycle */
+
+    @Override
+    void onBind() {
+        super.onBind();
+        bindLabel();
+        bindArrow();
+    }
+
+    @OnClick(R2.id.call_to_action_arrow)
+    void onTrigger() {
+        if (mCallbacks != null && (mModel == null || mModel.mEvents.isEmpty())) {
+            mCallbacks.goToNextPage();
+        } else if (mModel != null && !mModel.mEvents.isEmpty()) {
+            //TODO: trigger events
+        }
+    }
+
+    /* END lifecycle */
+
+    public void setCallbacks(@Nullable final Callbacks callbacks) {
+        mCallbacks = callbacks;
+    }
+
+    private void bindLabel() {
+        Text.bind(mModel != null ? mModel.mLabel : null, mLabelView);
+    }
+
+    private void bindArrow() {
+        final boolean visible = mModel == null || !mModel.getPage().isLastPage() || !mModel.mEvents.isEmpty();
+        mArrowView.setVisibility(visible ? View.VISIBLE : View.INVISIBLE);
+        mArrowView.setImageResource(R.drawable.ic_call_to_action);
+        mArrowView.setImageDrawable(DrawableUtils.tint(mArrowView.getDrawable(), CallToAction.getControlColor(mModel)));
+    }
+
+    @Override
+    protected void updateLayoutDirection() {
+        // force CallToAction to inherit it's layout direction
+        ViewCompat.setLayoutDirection(mRoot, ViewCompat.LAYOUT_DIRECTION_INHERIT);
+    }
+}
